# Use SimpleITK to resample