Fixes #271.
Resolves #227.

# Benchmark with Colin 1998 T1 (1 mm iso)

## Upsample (0.87 mm iso)
Old nearest: **2.35 s ± 360 ms**
Old linear: 2.9 s ± 567 ms
New nearest: **326 ms ± 13.8 ms**
New linear: 561 ms ± 233 ms

## Downsample (1.87 mm iso)
Old nearest: 236 ms ± 15.3 ms
Old linear: 335 ms ± 58.9 ms
New nearest: 236 ms ± 12.7 ms
New linear: 244 ms ± 12.1 ms