@fepegar 2020-08-17
@GFabien could you please take a look at `test_affine` in `test_resample.py`? I'm not sure it's correct.