@GFabien 2020-08-18
>@GFabien could you please take a look at test_affine in test_resample.py? I'm not sure it's correct.

I think it is correct, with the pre affine matrix you update the affine and resample using the updated affine. In the test, the affine is the identity, the pre affine matrix is
```python
np.array([
    [1,   0,   0,  10],
    [0,   1,   0,   0],
    [0,   0,  1.2,  0],
    [0,   0,   0,   1]
])
```
so the updated affine is equal to the pre affine matrix. The transform is supposed to resample the image to spacing `(1, 1, 1)` so we should end up with the following affine
```python
np.array([
    [1,   0,   0,  10],
    [0,   1,   0,   0],
    [0,   0,   1,   0],
    [0,   0,   0,   1]
])
```
or am I missing something? ITK finds the following one:
```python
np.array([
    [1,   0,   0,   10],
    [0,   1,   0,    0],
    [0,   0,   1, -0.1],
    [0,   0,   0,    1]
])
```