@fepegar 2020-08-18
Hmm I think I misinterpreted something in the test last night. Anyway, you're right.

However, a small displacement of the origin is expected if we want to keep exactly the same bounding box so I think the fixture should include that -0.1.

Think of a little square and a big square representing the first pixel (0, 0) of an image at 1 and 2 mm spacing. If both pixels are centered (image origin) at the same position, the bounds of the image will be different. To keep it constant, the small square needs to be a bit different. In this case, the difference is only in the z-axis because of the 1.2. Does any of that make sense?