@GFabien 2020-08-18
Yes it does, I think you're right. I just tried several spacing to check it and it seems to always generate a displacement of half the difference of spacings which is consistent with your statement. So yes, the fixture should take this -0.1 into account.