@fepegar 2020-08-18
FYI the origin displacement is computed in lines 208 and 209 of resample.py.

Thanks for checking that test! Maybe NiBabel doesn't change the origin and that's why it was passing. It's just an implementation detail, anyway.